django-backend/requirements_updated.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
bitarray==3.1.1
blinker==1.4
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
cytoolz==1.0.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
django-cors-headers==4.7.0
django-extensions==4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
exceptiongroup==1.2.2
fastapi==0.114.2
fonttools==4.53.0
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
hexbytes==1.3.0
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
more-itertools==8.10.0
multidict==6.1.0
narwhals==1.8.1
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
packaging==24.1
pandas==2.2.2
parsimonious==0.10.0
pillow==10.3.0
propcache==0.3.0
protobuf==5.28.1
psycopg2-binary==2.9.10
pyarrow==17.0.0
pycryptodome==3.21.0
pydantic==2.9.1
pydantic_core==2.23.3
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyunormalize==16.0.0
PyYAML==5.4.1
referencing==0.35.1
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
rich==13.8.1
rlp==4.1.0
rpds-py==0.20.0
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.38.5
streamlit==1.38.0
systemd-python==234
tenacity==8.5.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.1
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
uvicorn==0.30.6
wadllib==1.3.6
watchdog==4.0.2
web3==7.8.0
websockets==13.1
yarl==1.18.3
zipp==1.0.0
